+++ src/crypto/hash.js
