+++ src/crypto/sign.js
